renames VDI 6007 test cases #409
