+++ teaser/examples/verification/verification_VDI_6007_room10.py
