+++ teaser/examples/verification/verification_VDI_6007_room8.py
